add remarks on prescription request and resource;change column name

+++ app/Models/V1/Medicine/MedicinePrescription.php
@@ -38,7 +38,7 @@ class MedicinePrescription extends Model
 
     protected $casts = [
         'prescription_date' => 'date:Y-m-d',
-        'dosage_quantity' => 'integer',
+        'dosage_quantity' => 'double',
         'duration_intake' => 'integer',
         'quantity' => 'integer',
     ];
